MariaDBでrootログイン出来ない問題を修正

+++ databases/databases.sh
@@ -4,6 +4,10 @@ BASEDIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && cd . && pwd )"
 ENVFILE=$BASEDIR/.env
 eval $( cat $ENVFILE | perl -nlpE 's#^([^=]+)=(.+)$#export $1="$2"#' )
 
+PWFMA=$BASEDIR/.password-mariadb
+PWFPG=$BASEDIR/.password-pgsql
+
+
 # ボリュームの作成
 volume_create() {
 	local name=$1
@@ -16,57 +20,61 @@ volume_create() {
 
 # パスワードの作成
 create_password() {
-	local name=$1
-	local fn="$BASEDIR/.password-$name"
+	local fn=$1
 	if [ -e "$fn" ]; then
 		echo "$fn already exists."
 		exit 1
 	fi
-	perl -E 'my @chars; for(my $i=0;$i<$ARGV[1];$i++) { push @chars,substr($ARGV[0],int(rand()*length($ARGV[0])),1) }; say join("",@chars);' 'abcdefghjkmnpqrstuvwxyzABCDEFGHJKLMNPQRSTUVWXYZ23456789-_().!' 16 > $fn
+	perl -E 'my @chars; for(my $i=0;$i<$ARGV[1];$i++) { push @chars,substr($ARGV[0],int(rand()*length($ARGV[0])),1) }; say join("",@chars);' 'abcdefghjkmnpqrstuvwxyzABCDEFGHJKLMNPQRSTUVWXYZ23456789' 16 > $fn
 	chmod 600 $fn
 	echo "Create: $fn"
 }
 
 # MariaDBの初期化
 init_mariadb() {
-	create_password mariadb
-	local password=$(cat $BASEDIR/.password-mariadb)
-	docker-compose run -e MYSQL_ROOT_PASSWORD="$password" -d mariadb
+	create_password $PWFMA
+	local password=$(cat $PWFMA)
+	docker-compose run -e "MYSQL_ROOT_PASSWORD=$password" -d mariadb
 	local container=$(docker-compose ps -q mariadb)
 	perl - $container << 'EOS'
 use strict;
 $|=1;
 my $kid=open(my $log,"docker logs -f $ARGV[0] 2>&1 | ") || die $!;
+my $count=0;
 while(<$log>) {
+	print "$_";
 	if(m#\Qready for connections.\E#s ){
-		kill 2,$kid;
-		print "起動を確認しました\n";
-		exit(0);
+		$count++;
+		if($count==2) {
+			kill 2,$kid;
+			print "初期化を確認しました\n";
+			exit(0);
+		}
 	}
 	if(m#\Q[ERROR]\E#s ){
 		kill 2,$kid;
 		print "初期化に失敗しました\n";
 		exit(1);
 	}
-	print "$_"
 }
 EOS
 }
 
 # PostgreSQLの初期化
 init_pgsql() {
-	create_password pgsql
-	local password=$(cat $BASEDIR/.password-pgsql)
-	docker-compose run -e POSTGRES_PASSWORD="$password" -d pgsql
+	create_password $PWFPG
+	local password=$(cat $PWFPG)
+	docker-compose run -e "POSTGRES_PASSWORD=$password" -d pgsql
 	local container=$(docker-compose ps -q pgsql)
 	perl - $container << 'EOS'
 use strict;
 $|=1;
 my $kid=open(my $log,"docker logs -f $ARGV[0] 2>&1 | ") || die $!;
 while(<$log>) {
+	print "$_";
 	if(m#\QPostgreSQL init process complete; ready for start up.\E#s ){
 		kill 2,$kid;
-		print "起動を確認しました\n";
+		print "初期化を確認しました\n";
 		exit(0);
 	}
 	if(m#\Q[ERROR]\E#s ){
@@ -74,7 +82,6 @@ while(<$log>) {
 		print "初期化に失敗しました\n";
 		exit(1);
 	}
-	print "$_"
 }
 EOS
 }
@@ -88,6 +95,7 @@ destroy() {
 	rm -f $BASEDIR/.password-pgsql
 }
 
+# 作成
 do_create() {
 	destroy
 	volume_create $COMPOSE_PROJECT_NAME'-mariadb'
@@ -98,10 +106,18 @@ do_create() {
 }
 
 do_destroy() {
-	destory
+	destroy
 }
 
-COMMANDS="create destroy"
+do_mysql() {
+	exec docker-compose exec mariadb mysql -p$(cat $PWFMA) $@
+}
+
+do_psql() {
+	exec docker-compose exec pgsql psql -U postgres $@
+}
+
+COMMANDS="create destroy mysql psql"
 
 function run {
 
@@ -121,4 +137,3 @@ function run {
 }
 
 run $@
-
